Add handling for registering preset block variations
This is maybe a little weird, and may need more work once we start
adding additional presets.

To use it, set the default block registration as a `const`, and pass
it along with a defined template to the `RegisterBlockPreset` function,
which registers a modified copy of the default block. Most notably,
the new block is registered with a `presetTemplate` property that is
passed to the `template` property of the `InnerBlocks` component
in the `edit` function.

Because `presetTemplate` isn't defined in the default block, the
`template` property is simply ignored in that context.