Generate an index-based value for the anchor attribute if it's not set
This is just a baby step - we'll somehow need to account for the
following scenario:

If there are ten headlines on a post, headline two is deleted, and a
new headline is added after the former tenth headline, that and the
newly-added headline with both have the same id.

This also doesn't account for heading levels - the index is just
based on all headline blocks, so we'll probably want to get more
sophistcated with that, too.

But the cool thing here is that it allows for overriding the generated
id via the HTML Anchor field :)