Indicate when a modal or its innerblocks are selected
This adds a `data-selected-modal` attribute to the edit wrapper of
the modal block, which will be leveraged in the CSS for keeping the
modal content component visible as needed.

This is handled in the `getEditWrapperProps` function so we can get
an accurate result from `getSelectedBlockClientId()`, which would only
return while the component block is selected if it were inside the
`edit` function.

It's probably not the most efficient. I'm going to call it good for
now, though, since I've already had too much fun with it.