Update background opacity controls and output classes
This defaults the control value to 100 and only outputs classes
on the element if the value is set to something below that. It also
sets the minimum value to 10, because going completely transparent
doesn't make a lot of sense.

I've also added some CSS for displaying these changes - a SASS
wizard I am not, so there may well be a more efficient way to write
those rules out.