Add publication colors to the default Gutenberg palette
This isn't quite what I had in mind when I started down this road,
but Gutenberg returns the full color object - including the name/
class, which is specifically what we're after here - if it's set
in the editorSettings > colors object. (Otherwise it just returns
the hex code.) I suspect this is how it handles colors from the
picker, so I'll be keeping an eye out for ways to improve this.