Set editable area to the same width as the placeholder text
This felt confusing before, as the cursor would be centered in the
block instead of over the placeholder. Although this change is more
in line with the default Gutenberg button experience, it still seems
a little strange as it shrinks the target area. We'll see how it
tests and adjust accordingly :).