Refine layout options
This adds a few options I missed previously, adds styles for displaying
images as the option labels, and organizes the options in a way that I
think makes sense :).

We seem to be missing images for the `photo-row-fourths-2-2` and
`photo-row-short-2-2-2-2` options, which I believe are supported
as they are represented in the styleguide.