Add images and styles for the video input placeholder
We might consider changing these images out for higher-quality
versions - I just snipped these using Preview :).

I opted to use divs in the JSX and apply the images as backgrounds
to avoid having to deal with loader modules and the funkiness that
can occur when using images as flex objects.