Adjust URL parameters in attempt to get autoplay working
This combo of parameters and attributes worked on a regular YouTube
embed when I was working on the background component, but no dice
here. I'm committing this anyway as I think it makes the build out
of the full URL more readable and doesn't add attributes if they're
not needed.