Add styles and functions for improving the item number control
Gutenberg does some interesting things with inline heights on these
textareas, so the `min-` and `max-height` rules are in place to keep
that in check.

I don't fully understand what's going on with the width, but I wasn't
able to make it nice through just CSS. So I've added an inline width
to the control that is either 100% when it's empty or `{n}ch`, where
`n` is the number of characters in the input.