Render the Featured Story block in PHP
This change is so that we can get an accurate and current comment
count when outputting this block. Evidently, dynamic block attributes
can't be sourced from the output (which kind of makes sense) - so this
required some adjustments to not only the Featured Story attributes,
but the `head` and `deck` attributes on the Leading block too.