Match expected attribute type on custom HTML block
The parser used to check blocks for existing attributes is more
strict than I thought. https://github.com/aduth/hpq

If you specify the type of an attribute as a `string`, and then
provide an `integer` (see `new Date().time`) as the attribute, then
it will be stripped out on subsequent page loads.

With the scripted block insertions, I used `banner-` and `migrated-`
to classify how they appeared in content. I apparently did not do
enough testing with blocks that were manually added.

By prepending `manual-` to generated block IDs, the ID becomes a
string and blocks save and reappear in saved content as expected.