:fire: Replace AaronHolbrook\Autoload with require
Due to performance reasons, autoload.php from the vendor folder will be removed. Prior to this, it was used to load all PHP files recursively in the /src/ directory so developers can focus on creating blocks without having to use include_once or require_once. This became a performance issue post launch that was brought up by webteam with NewRelic stats on the most intensive functions used in the One Ed project.

Manually including should resolve that performance issue.