fix bug in loop/event handling for multiple blocks
If more than 1 drawer block or more than 1 modal block is on the page the old for loop couldn’t work to properly assign the event handlers to the right element. Clicking the first block caused the 2nd block to open. This has to do wih how for… loops work. Changing it to a forEach lets it work as expected. Also switched some selectors to use QuerySelector