Use post date as pre-formatted in related stories list
I originally used `__experimentalGetSettings().formats.date` as a
way of retrieving the expected format in which the date should be
displayed. It seems that I missed the date was already being
provided through the post data in the expected format, so the use
of `format()` and `__experimentalGetSettings()` is not necessary here.