Only call `onChange` if it exists as a function
In the latest version of WordPress and Gutenberg, it's possible
the `onChange` function does not exist by default on block
components.

This change checks that `onChange` is callable before attempting to
call it.

This will likely be fixed in future versionso of Gutenberg via
https://github.com/WordPress/gutenberg/pull/17036, but it won't
hurt to have a fallback check here.