Use a dynamic template for photoessay blocks
Rather than watch for changes and insert/remove blocks depending
on the layout, we can use a template to enforce the number of blocks
that appear in the rendered component.

This works in 5.x with the latest Gutenberg plugin, but has not yet
been tested in 4.9.

New blocks do not yet immediately receive the class name.