Only use editor settings when block level settings are unavailable
In WordPress 5.2, several pieces of Gutenberg were split into block
editor components as a way of indicating they could be used outside
of the standard editor. See #177 for initial work on this repo to
adapt to that.

This commit adds some fallbacks to use `getEditorSettings()` and
`updateEditorSettings()` as normal when `getSettings()` and
`updateSettings()` are not yet available.