limit formats to headline block only. Add new icons to toolbar
After the update branches were merged the PreText and PostText formats for the headline block were appearing on all RichText toolbars on all blocks (edition, article blocks, etc). This commit adds a withSelect function and ifCondition to check if the selected block is `editorial/headline` and only show the formats in the richtext toolbar if that block is selected and active in the editor.

In addition custom SVG icons have been added to replace the placeholder `warning` icon for these two formats.