update class names for 5.x’s new block editor naming conventions
at some point in the 5.x WP releases the gutenberg editor’s classnames have been changed to start with `.block-editor` instead of `.editor-` and this is breaking UI elements and block previews throughout the Gutenberg editor.

This commit should add these new class names to the styles for the 4.9 version of gutenberg so it should work in both 4.9 of WP and newer versions.