deprecate past version of the block
We need to deprecate the past version of the block so we can add a new class `wp-block-bu-button` as `wp-block-button` was a poor choice originally since WP uses that class. We can’t remove it entirely but adding a new secondary class should let us start the process of changing this block to be more bu specific.