the url input does not render properly in WP 5.4
This looks like it was a hack to recreate a url input component in the early version of the gutenberg plugin.

I haven’t found an example for adding a link button to the block toolbar which is where it should be for this button block so a quick temporary fix is to move the link field to the inspector panel (sidebar) and remove it from the editor view.

This is copied from the example 10up added to the Leadin Block.

ToDo: in the future this url input should move to the toolbar (but not as a formatting option).