refactor how block is saved
The block is not saving properly in 5.4 (and in 4.9 but to less of a degree).

Saving the post meta when the block saves to the post_content is the wrong approach I believe. The save function in a Gutenberg block is NOT === to the saving of a post. A user may edit the block and change the HTML but not save the post itself but in WP 4.9 those changes are saved to post meta regardless.

In 5.4 it does the same thing, but less reliably. Something with the `states` might have changed, I haven’t been able to identify it. But regardless this change to use `wp.data.subscribe` and monitor the the saving of the post in the block editor I think is a safer approach. It appears to save reliably in my testing.

However I’m not sure if this is the best way for the future. This block is a bit outside the norm so I haven’t found great reference guides for handling block data for something like this.