new heading icon fuction
The previous approach to create heading icons using a single heading icon and appended a subscript text with the heading number does not render properly in 5.4. Instead of using css hacks to force it to work I found this function from gutenberg core for the existing heading block and reused it in our block.