dropcap allows original size images, this tries to avoid that
This enhancement checks for certain image sizes and if available will use those before falling back to the original full size image.

Phil had pointed out that there are cases with the block editor in certain conditions where media.sizes are not available yet so these conditional checks are safer. I’m not sure if that has been mitigated in later versions of Gutenberg or not so I’m keeping with that practice.